+++ LumenSys.WebAPI/Controllers/ContractsController.cs
@@ -1,0 +1,6 @@
+﻿namespace LumenSys.WebAPI.Controllers
+{
+    public class ContractsController
+    {
+    }
+}
