+++ LumenSys.WebAPI/Data/Builders/ContractsBulder.cs
@@ -1,0 +1,6 @@
+﻿namespace LumenSys.WebAPI.Data.Builders
+{
+    public class ContractsBulder
+    {
+    }
+}
