+++ LumenSys.WebAPI/Objects/DTOs/Entities/ContractsDTO.cs
@@ -1,0 +1,6 @@
+﻿namespace LumenSys.WebAPI.Objects.DTOs.Entities
+{
+    public class ContractsDTO
+    {
+    }
+}
