+++ LumenSys.WebAPI/Services/Entities/ContractsService.cs
@@ -1,0 +1,6 @@
+﻿namespace LumenSys.WebAPI.Services.Entities
+{
+    public class Contracts
+    {
+    }
+}
