+++ LumenSys.WebAPI/Services/Interfaces/IContractsService.cs
@@ -1,0 +1,6 @@
+﻿namespace LumenSys.WebAPI.Services.Interfaces
+{
+    public interface IContractsService
+    {
+    }
+}
